JP Onboarding: Move data layer bits to appropriate location
